Insert new abstract class which contains common code for DistSetDetails
